// lib/swedish-regions.generated.ts
export type SwedishCounty = {
  code: string
  name: string
  slug: string
}

export type SwedishMunicipality = {
  code: string
  name: string
  slug: string
  countyCode: string
  county: string
  countySlug: string
}

export const swedishCounties: SwedishCounty[] = [
  {
    "code": "01",
    "name": "Stockholms län",
    "slug": "stockholms-lan"
  },
  {
    "code": "03",
    "name": "Uppsala län",
    "slug": "uppsala-lan"
  },
  {
    "code": "04",
    "name": "Södermanlands län",
    "slug": "sodermanlands-lan"
  },
  {
    "code": "05",
    "name": "Östergötlands län",
    "slug": "ostergotlands-lan"
  },
  {
    "code": "06",
    "name": "Jönköpings län",
    "slug": "jonkopings-lan"
  },
  {
    "code": "07",
    "name": "Kronobergs län",
    "slug": "kronobergs-lan"
  },
  {
    "code": "08",
    "name": "Kalmar län",
    "slug": "kalmar-lan"
  },
  {
    "code": "09",
    "name": "Gotlands län",
    "slug": "gotlands-lan"
  },
  {
    "code": "10",
    "name": "Blekinge län",
    "slug": "blekinge-lan"
  },
  {
    "code": "12",
    "name": "Skåne län",
    "slug": "skane-lan"
  },
  {
    "code": "13",
    "name": "Hallands län",
    "slug": "hallands-lan"
  },
  {
    "code": "14",
    "name": "Västra Götalands län",
    "slug": "vastra-gotalands-lan"
  },
  {
    "code": "17",
    "name": "Värmlands län",
    "slug": "varmlands-lan"
  },
  {
    "code": "18",
    "name": "Örebro län",
    "slug": "orebro-lan"
  },
  {
    "code": "19",
    "name": "Västmanlands län",
    "slug": "vastmanlands-lan"
  },
  {
    "code": "20",
    "name": "Dalarnas län",
    "slug": "dalarnas-lan"
  },
  {
    "code": "21",
    "name": "Gävleborgs län",
    "slug": "gavleborgs-lan"
  },
  {
    "code": "22",
    "name": "Västernorrlands län",
    "slug": "vasternorrlands-lan"
  },
  {
    "code": "23",
    "name": "Jämtlands län",
    "slug": "jamtlands-lan"
  },
  {
    "code": "24",
    "name": "Västerbottens län",
    "slug": "vasterbottens-lan"
  },
  {
    "code": "25",
    "name": "Norrbottens län",
    "slug": "norrbottens-lan"
  }
]